+++ android/app/src/main/AndroidManifest.xml
@@ -6,7 +6,8 @@
     <!-- Permissions for saving PDFs and accessing storage -->
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
-
+    <uses-permission android:name="android.permission.MANAGE_EXTERNAL_STORAGE" />
+    
     <application
         android:label="school_erp"
         android:name="${applicationName}"
