+++ lib/screens/Teacher_screens.dart/teacher_home_screen.dart
@@ -219,7 +219,7 @@ class _TeacherHomeScreenState extends State<TeacherHomeScreen> {
                               onTap: () => Navigator.push(
                                 context,
                                 MaterialPageRoute(
-                                  builder: (_) => const EventsScreen(),
+                                  builder: (_) => EventsScreen(),
                                 ),
                               ),
                             ),
