+++ lib/screens/events/event.dart
@@ -4,21 +4,21 @@ class Event {
   final String title;
   final String description;
   final DateTime date;
-  final String? imageUrl; // Add this line
+  final List<String>? mediaUrls; // Assume mediaUrls is a list
 
   Event({
     required this.title,
     required this.description,
     required this.date,
-    this.imageUrl, // Add this line
+    this.mediaUrls,
   });
 
   factory Event.fromMap(Map<String, dynamic> data) {
     return Event(
-      title: data['title'],
-      description: data['description'],
+      title: data['title'] ?? '',
+      description: data['description'] ?? '',
       date: (data['date'] as Timestamp).toDate(),
-      imageUrl: data['imageUrl'], // Add this line
+      mediaUrls: List<String>.from(data['mediaUrls'] ?? []), // List of URLs
     );
   }
 }
