+++ lib/screens/events/events_screen.dart
@@ -1,87 +1,15 @@
-import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/material.dart';
+import 'package:school_erp/screens/events/event.dart';
+import 'package:school_erp/screens/events/event_detail_screen.dart';
 
-// Event model (same as before)
-class Event {
-  final String title;
-  final String description;
-  final DateTime date;
-  final String? imageUrl;
-
-  Event(
-      {required this.title,
-      required this.description,
-      required this.date,
-      this.imageUrl});
-
-  factory Event.fromMap(Map<String, dynamic> data) {
-    return Event(
-      title: data['title'],
-      description: data['description'],
-      date: (data['date'] as Timestamp).toDate(),
-      imageUrl: data['imageUrl'],
-    );
-  }
-}
-
-// EventDetailScreen
-class EventDetailScreen extends StatelessWidget {
-  final Event event;
-
-  const EventDetailScreen({Key? key, required this.event}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(event.title),
-      ),
-      body: SingleChildScrollView(
-        padding: EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (event.imageUrl != null && event.imageUrl!.isNotEmpty)
-              Center(
-                child: Image.network(
-                  event.imageUrl!,
-                  height: 200,
-                  fit: BoxFit.cover,
-                ),
-              ),
-            SizedBox(height: 16),
-            Text(
-              event.title,
-              style: Theme.of(context).textTheme.headlineSmall,
-            ),
-            SizedBox(height: 8),
-            Text(
-              "${event.date.toLocal()}".split(' ')[0],
-              style: TextStyle(fontSize: 14, color: Colors.grey),
-            ),
-            SizedBox(height: 16),
-            Text(
-              event.description,
-              style: Theme.of(context).textTheme.bodyMedium,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// Updated EventsScreen
 class EventsScreen extends StatefulWidget {
-  const EventsScreen({Key? key}) : super(key: key);
-
   @override
-  State<EventsScreen> createState() => _EventsScreenState();
+  _EventsScreenState createState() => _EventsScreenState();
 }
 
 class _EventsScreenState extends State<EventsScreen> {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
-  late Future<List<Event>> _events;
 
   Future<List<Event>> fetchEvents() async {
     QuerySnapshot snapshot = await _firestore.collection('events').get();
@@ -91,19 +19,13 @@ class _EventsScreenState extends State<EventsScreen> {
   }
 
   @override
-  void initState() {
-    super.initState();
-    _events = fetchEvents();
-  }
-
-  @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text("Events"),
       ),
       body: FutureBuilder<List<Event>>(
-        future: _events,
+        future: fetchEvents(),
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
             return Center(child: CircularProgressIndicator());
@@ -117,12 +39,6 @@ class _EventsScreenState extends State<EventsScreen> {
               itemBuilder: (context, index) {
                 final event = snapshot.data![index];
                 return ListTile(
-                  leading: event.imageUrl != null && event.imageUrl!.isNotEmpty
-                      ? CircleAvatar(
-                          backgroundImage: NetworkImage(event.imageUrl!),
-                          radius: 25,
-                        )
-                      : Icon(Icons.event, size: 50),
                   title: Text(event.title),
                   subtitle: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
