+++ lib/screens/student_screens/home_screen.dart
@@ -237,7 +237,7 @@ class _HomeScreenState extends State<HomeScreen> {
                               onTap: () => Navigator.push(
                                 context,
                                 MaterialPageRoute(
-                                  builder: (_) => const EventsScreen(),
+                                  builder: (_) => EventsScreen(),
                                 ),
                               ),
                             ),
